Search_Word_Meaning.py
Take  a Word as input from user. If the word is valid, the meaning/meanings  is/are displayed.
If entered word is invalid,  similar word is displayed and user is asked for yes/ no (if the displayed word is the desired word).
If not,then no result is shown and user is asked to enter valid word again for searching.